Update README.md
1.1 Debugging Assembly

7.1 For information on debugging or the autograder, jump to the debugging section. Please note that Docker must be running, but it doesn’t need to be within the image.

1.2 Factorial

Given a value n, calculate the factorial by implementing the Factorial and Multiply Subroutines. Return the final answer. The pseudocode can be found in section 3.1.1.

Below is a list of what we have provided you for this method.

    label named “MULTIPLY” where the multiply subroutine begins

    label named “FACTORIAL” where the factorial subroutine begins

    label named “STACK” where the start of the stack is located

1.3 Insertion Sort

Given an array of integers, you will be tasked with implementing an insertion sort to put them in ascending order. The pseudocode can be found in section 3.2.1.

Below is a list of what we have provided you for this method.

    label named “INSERTION SORT” where the insertion sort subroutine begins

    label named “STACK” where the start of the stack is located

    values of the original array starting at x4000